backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
asyncpg==0.30.0
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
passlib[bcrypt]==1.7.4
PyJWT==2.9.0
boto3==1.35.36
python-multipart==0.0.9
